feat(utils): hide mapbox token on serverside and expose temporal mapboxTokens to clientside
