+++ .env.template
@@ -1,1 +1,2 @@
-NEXT_PUBLIC_MAPBOX_TOKEN = 
+MAPBOX_TOKEN_DEV = 
+MAPBOX_SECRET_PROD_TOKEN =
