+++ app/page.tsx
@@ -1,9 +1,12 @@
-"use client";
 import Image from "next/image";
+
+import getMapboxToken from "@/utils/getMapboxToken";
 
 import LandingSearch from "./components/LandingSearch";
 
-export default function Home() {
+export default async function Home() {
+  const mapboxToken = await getMapboxToken();
+
   return (
     <section className="flex max-h-screen flex-col items-center justify-between p-24">
       <div className="relative flex place-items-center before:absolute before:h-[300px] before:w-[480px] before:-translate-x-1/2 before:rounded-full before:from-white before:to-transparent before:content-[''] after:absolute after:-z-20 after:h-[180px] after:w-[240px] after:translate-x-1/3 after:bg-gradient-conic after:from-sky-200 after:via-blue-200 after:blur-2xl after:content-[''] before:dark:bg-gradient-to-br before:dark:from-transparent before:dark:to-blue-700 before:dark:opacity-10 after:dark:from-sky-900 after:dark:via-[#0141ff] after:dark:opacity-40 before:lg:h-[360px]">
@@ -17,7 +20,7 @@ export default function Home() {
         />
       </div>
       <section>
-        <LandingSearch />
+        <LandingSearch mapboxToken={mapboxToken} />
       </section>
     </section>
   );
