+++ utils/getGeocoder.tsx
@@ -2,9 +2,7 @@ import MapboxGeocoder from "@mapbox/mapbox-gl-geocoder";
 import Mapbox from "mapbox-gl";
 
 import geojson from "../data/places.json";
-export default function getGeocoder(): MapboxGeocoder {
-  const MAPBOX_TOKEN = process.env.NEXT_PUBLIC_MAPBOX_TOKEN; // Set your mapbox token here
-
+export default function getGeocoder(mapboxToken: string): MapboxGeocoder {
   function forwardGeocoder(query: any) {
     const matchingFeatures = [];
     for (const feature of geojson.features) {
@@ -25,7 +23,7 @@ export default function getGeocoder(): MapboxGeocoder {
   }
 
   return new MapboxGeocoder({
-    accessToken: MAPBOX_TOKEN as string,
+    accessToken: mapboxToken as string,
     localGeocoder: forwardGeocoder,
     localGeocoderOnly: true,
     mapboxgl: Mapbox,
